migration内の型を再度修正
浮動小数点数（float、double）
固定小数点（decimal）
※固定小数点を採用。
※浮動小数点数だと、値をDBに保存する時に10進法での入力値がそのまま入らない事があり、かつ合計した場合のずれがどの程度まで膨れるかわからない為
。

+++ database/migrations/2019_08_20_203450_create_hours_table.php
@@ -16,7 +16,7 @@ class CreateHoursTable extends Migration
         Schema::create('hours', function (Blueprint $table) {
             $table->increments('id');
             $table->date('date');
-            $table->double('hour', 5, 3 );
+            $table->decimal('hour', 5, 3 );
                 // 要確認！
             $table->unsignedInteger('category_id');
             $table->unsignedInteger('task_id')->nullable();
